Cleaned and all codes and outputs
The last commit has all the codes, final sets of data as used in the codes, and outputs. The binary results are giving good results. Note: when Lasso is made to work on EC as a factor vs. numeric, some additional parameters get dropped. This is running EC as a factor. Also, in the continuous EC, ANN seems to be giving different scaled results and quantile regression is not working. Left those for now.